Fix open files limit in forked child process
There has been a change in how Go 1.21 handles the restoration of the NOFILE resource rlimit in child processes.
To address this change and ensure that the NOFILE rlimit is properly set in the Envoy process,
before forking the Envoy process, we verify and configure the NOFILE rlimit correctly in the parent process.
This ensure that the child process inherits the same resource limits.